separate arch dependent versions of m2c-tokenset.c
C is one hell of a sucky language. The C99 standard mandates that the bitwidths and largest values of the long long types are defined as macros in limits.h. Alas, these macros are not defined on several systems as limits.h appears to be overriden by the OS environment or configuration or compiler installation. So here we go, separate implementation files, one for each target architecture, 16, 32 and 64 bit. C doesn't do cross-platform portability anywhere near as well as it is made out to be. Not having constants in the language is part of the problem here. Fools gold.