production rules renamed
Due to the inherited but incorrect nomenclature "definition module" for interface modules, we had redefined the terms definition and declaration to mean the exact opposite of what they actually mean. This wasn't really tenable and has now been corrected by changing the syntax to INTERFACE and the terminology back to definition=definition and declaration=declaration. Consequently, the grammar rules had to be changed accordingly.